edit product info dialogs

+++ app/src/main/java/comjason_lewisg/httpsgithub/boozic/Models/UpdateProductModel.java
@@ -47,7 +47,7 @@ public class UpdateProductModel {
         updated = true;
     }
 
-    public void updateVolume(int volume) {
+    public void updateVolume(double volume) {
         this.volume = volume;
         updated = true;
     }
